requirements.txt
streamlit
torch
torchvision
pillow
google-cloud-storage